requirements.txt
certifi==2024.2.2
elastic-transport==8.13.0
elasticsearch==7.17
gensim==4.3.2
joblib==1.4.0
numpy==1.26.4
scipy==1.10.1  # 'triu' from 'scipy.linalg'
smart-open==7.0.4
urllib3==2.2.1
wrapt==1.16.0
mecab-python3==1.0.8